Hi all,

This PR is to add support for Cisco ASAv, I've build and tested this with the asav9-18-2.qcow2 image that is available from Cisco CML / VIRL.

I'd like to add it to this project so it hopefully can also be supported in https://github.com/srl-labs/containerlab

Contact me if you need help building it @plajjan